+++ backend/cwa_lib/pages/su_manage_doc_tasks.py
@@ -1,13 +1,12 @@
 from common import log
 from common.sql_db_async import AsyncSession
 from common.sql_models import DocTasks
-from common.sql_tools import create_order_clause
-from sqlalchemy import select, delete
+from sqlalchemy import delete
 from cwa_lib.pydantic_schemas.generic_table import (
-    ColumnType, TableOptions, TableQuery, TableDeleteRowResult
+    ColumnType, TableOptions, TableDeleteRowResult
 )
-from cwa_lib.pages import get_qc_to_with_user_id_name_group_id_name
-from cwa_lib.pydantic_schemas.su_manage_doc_tasks import SuManageDocTasksQueryResult
+from cwa_lib.pydantic_schemas.su_manage_doc_tasks import SuManageDocTaskRead
+from cwa_lib.pages import GenericTableRead
 
 su_manage_doc_tasks__query_columns = {
     'doc_task_id': ColumnType(display='ID', type='number'),
@@ -46,46 +45,26 @@ su_manage_doc_tasks__table_options = TableOptions(
     order_by__allow=['doc_task_id', 'group_id', 'user_id', 'status', 'created_at']
 )
 
+class SuManageDocTasksTableRead(GenericTableRead):
+    sa_model = DocTasks
+    read_model = SuManageDocTaskRead
+    name = 'su_manage_doc_tasks'
+    query_columns = su_manage_doc_tasks__query_columns
+    table_options = su_manage_doc_tasks__table_options
+    default_order_by = table_options.pk
+    qc_to_user_group = {'user_id': ('add_values', 'select_default', 'allow_all'), 'group_id': ('add_values', 'select_default', 'allow_all')}
+
+    def _get_where_clause(self):
+        where_clause = DocTasks.doc_task_id > -1
+        if (deleted := self.kwargs.get('deleted', 0)) is not None:
+            where_clause &= DocTasks.deleted == deleted
+        return where_clause
+
 
 class SuManageDocTasksTable:
     def __init__(self, session: AsyncSession) -> None:
         self.session = session
 
-    async def query_all(
-            self,
-            payload: TableQuery,
-            deleted: int | None
-            ) -> SuManageDocTasksQueryResult:
-        # get values to make 'User ID: Name', 'Group ID: Name' columns
-        manage_doc_tasks__qc, manage_doc_tasks__to = await get_qc_to_with_user_id_name_group_id_name(
-            self.session, su_manage_doc_tasks__query_columns, su_manage_doc_tasks__table_options,
-            {'user_id': ('add_values', 'select_default'), 'group_id': ('add_values', 'select_default')}
-        )
-        #
-        where_clause = DocTasks.doc_task_id > -1
-        if deleted is not None:
-            where_clause &= DocTasks.deleted == deleted
-        order_clause, order_by, order_dir = create_order_clause(
-            DocTasks, manage_doc_tasks__to.pk, payload.order_by, payload.order_dir
-        )
-        result = await self.session.execute(
-            select(DocTasks)
-            .where(where_clause)
-            .order_by(order_clause)
-            .limit(payload.limit)
-            .offset(payload.offset)
-        )
-        rows = result.scalars().all()
-        return SuManageDocTasksQueryResult(
-            name='su_manage_doc_tasks',
-            rows=rows,
-            columns=manage_doc_tasks__qc,
-            table_options=manage_doc_tasks__to,
-            order_by=order_by,
-            order_dir=order_dir,
-            total=len(rows)
-        )
-    
     async def delete_by_doc_task_id(self, doc_task_id: int) -> TableDeleteRowResult:
         """
         Delete one by doc_task_id