+++ backend/cwa_lib/routers/su_manage_log_crud.py
@@ -4,7 +4,7 @@ from common.sql_models.api_users import User
 from cwa_lib.app import current_active_user
 from cwa_lib.pydantic_schemas.generic_table import TableQuery, TableDeleteRowResult
 from cwa_lib.pydantic_schemas.su_manage_log_crud import SuManageLogCRUDQueryResult
-from cwa_lib.pages.su_manage_log_crud import SuManageLogCRUDTable
+from cwa_lib.pages.su_manage_log_crud import SuManageLogCRUDTableRead, SuManageLogCRUDTable
 
 router__su_manage_log_crud = APIRouter()
 
@@ -17,7 +17,7 @@ async def su_manage_log_crud__query(
     ):
     if not user.is_superuser:
         raise HTTPException(status_code=404, detail="Not found")
-    result = await SuManageLogCRUDTable(session).query_all(payload=payload)
+    result = await SuManageLogCRUDTableRead(session, payload).query()
     return result
 
 @router__su_manage_log_crud.delete("/su/manage_log_crud/{lc_id:int}", response_model=TableDeleteRowResult)