+++ backend/cwa_lib/routers/su_manage_users.py
@@ -4,7 +4,7 @@ from common.sql_models.api_users import User
 from cwa_lib.app import current_active_user
 from cwa_lib.pydantic_schemas.generic_table import TableQuery, TableCreateRowResult, TableUpdateRowResult, TableDeleteRowResult
 from cwa_lib.pydantic_schemas.su_manage_users import SuManageUsersQueryResult, SuManageUsersCreate, SuManageUsersUpdate
-from cwa_lib.pages.su_manage_users import SuManageUsersTable
+from cwa_lib.pages.su_manage_users import SuManageUsersTableRead, SuManageUsersTable
 from cwa_lib.sql_tables.log_crud import LogCRUDTable
 
 router__su_manage_users = APIRouter()
@@ -19,7 +19,7 @@ async def su_manage_users__query(
     if not user.is_superuser:
         raise HTTPException(404, detail="Not found")
 
-    result = await SuManageUsersTable(session).query_all(payload=payload, deleted=0)
+    result = await SuManageUsersTableRead(session, payload, deleted=0).query()
     return result
 
 @router__su_manage_users.post("/su/manage_users", response_model=TableCreateRowResult)